# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: JanVoice AI - Civic Engagement Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already in use
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
  /auth/me/profile:
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ── CAMPAIGNS ─────────────────────────────────────────────────────────────
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List all campaigns
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignList"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a new campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
...
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get campaign by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}/join:
    post:
      operationId: joinCampaign
      tags: [campaigns]
      summary: Join a campaign as a supporter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Joined successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /campaigns/{id}/bookmark:
    post:
      operationId: bookmarkCampaign
      tags: [campaigns]
      summary: Bookmark or unbookmark a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bookmark toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /campaigns/{id}/updates:
    get:
      operationId: getCampaignUpdates
      tags: [campaigns]
      summary: Get updates for a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign updates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CampaignUpdate2"
    post:
      operationId: postCampaignUpdate
      tags: [campaigns]
      summary: Post an update to a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdateInput"
      responses:
        "201":
          description: Update posted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignUpdate2"

  # ── PETITIONS ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
        password:
          type: string
          minLength: 8
        state:
          type: string
        city:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [guest, citizen, volunteer, moderator, admin, super_admin]
        avatarUrl:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        state:
          type: string
        city:
          type: string
        avatarUrl:
          type: string

    # ── CAMPAIGN SCHEMAS ─────────────────────────────────────────────────
    Campaign:
      type: object
      required: [id, title, description, category, status, supporterCount, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [education, infrastructure, healthcare, environment, public_services, women_safety, youth, other]
        status:
          type: string
          enum: [active, completed, pending, archived]
        imageUrl:
          type: ["string", "null"]
        supporterCount:
          type: integer
        signatureCount:
          type: integer
        targetSignatures:
          type: ["integer", "null"]
        progressPercent:
          type: integer
        isBookmarked:
          type: boolean
        isJoined:
          type: boolean
        location:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        creatorId:
          type: integer
        creatorName:
          type: string
        createdAt:
          type: string
          format: date-time
        endsAt:
          type: ["string", "null"]
    CampaignList:
      type: object
      required: [campaigns, total, page, limit]
      properties:
        campaigns:
          type: array
          items:
            $ref: "#/components/schemas/Campaign"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    CampaignInput:
      type: object
      required: [title, description, category]
      properties:
        title:
          type: string
          minLength: 5
        description:
          type: string
          minLength: 20
        category:
          type: string
        imageUrl:
          type: string
        location:
          type: string
        tags:
          type: array
          items:
            type: string
        targetSignatures:
          type: integer
        endsAt:
          type: string
          format: date-time
    CampaignUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
        imageUrl:
          type: string
        location:
          type: string
        tags:
          type: array
          items:
            type: string
    CampaignUpdate2:
      type: object
      required: [id, campaignId, title, content, createdAt]
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        title:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    CampaignUpdateInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string

    # ── PETITION SCHEMAS ─────────────────────────────────────────────────
...
    ActionResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string