# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * JanVoice AI - Civic Engagement Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  bio?: string | null;
  createdAt: Date;
}